requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
python-multipart==0.0.20
aiohttp==3.12.6
pyjwt==2.10.1
pydantic-settings==2.9.1
python-docx==1.1.2
openpyxl==3.1.5
